+++ src/main/java/com/github/hcsp/inheritance/Cat.java
@@ -1,15 +1,11 @@
 package com.github.hcsp.inheritance;
 
-public class Cat {
-    private String name;
+public class Cat extends Animal {
 
     public Cat(String name) {
-        this.name = name;
+        super(name);
     }
 
-    public void sayMyName() {
-        System.out.println("我的名字是" + name);
-    }
 
     public void meow() {
         System.out.println("喵");
