+++ src/main/java/com/github/hcsp/inheritance/Person.java
@@ -1,0 +1,13 @@
+package com.github.hcsp.inheritance;
+
+public class Person extends Animal{
+    private int age;
+    public void sayMyAge() {
+        System.out.println("我的年龄是" + age);
+    }
+
+    public Person(String name, int age) {
+        super(name);
+        this.age = age;
+    }
+}
